image-story/requirements.txt
transformers
langchain-community 
langchain-huggingface
faiss-cpu 
tiktoken 
python-dotenv
requests
diffusers